+++ pages/api/sell/[id].ts
@@ -6,16 +6,21 @@ export default async function handler(
   req: NextApiRequest,
   res: NextApiResponse<any>
 ) {
-  if (req.method == "PUT") {
-    const sell = await putSell(req)
-    res.status(201).json(sell)
-  } else if (req.method == "DELETE") {
-    const sell = await deleteSell(req)
-    res.status(201).json(sell)
-  } else {
-    let sell = await getSell(req)
-    res.status(200).json(sell)
-  }
+  if (req.headers.authorization !== process.env.API_KEY) {
+    return res.status(401).send("You are not authorized to call this API");
+  } else{ 
+    if (req.method == "PUT") {
+      const sell = await putSell(req)
+      res.status(201).json(sell)
+    } else if (req.method == "DELETE") {
+      const sell = await deleteSell(req)
+      res.status(201).json(sell)
+    } else {
+      let sell = await getSell(req)
+      res.status(200).json(sell)
+    }
+   }
+  
 }
 async function putSell(req: NextApiRequest): Promise<Array<Sell>> {
   const query:string = req.query.id as string;
