+++ pages/api/user/[id].ts
@@ -19,16 +19,21 @@ export default async function handler(
   req: NextApiRequest,
   res: NextApiResponse<any>
 ) {
-  if (req.method == "PUT") {
-    const user = await putUser(req)
-    res.status(201).json(user)
-  } else if (req.method == "DELETE") {
-    const user = await deleteUser(req)
-    res.status(201).json(user)
-  } else {
-    let user = await getUser(req)
-    res.status(200).json(user)
-  }
+  if (req.headers.authorization !== process.env.API_KEY) {
+    return res.status(401).send("You are not authorized to call this API");
+  } else{ 
+    if (req.method == "PUT") {
+      const user = await putUser(req)
+      res.status(201).json(user)
+    } else if (req.method == "DELETE") {
+      const user = await deleteUser(req)
+      res.status(201).json(user)
+    } else {
+      let user = await getUser(req)
+      res.status(200).json(user)
+    }
+   }
+  
 }
 async function putUser(req: NextApiRequest): Promise<Array<User>> {
   const query:string = req.query.id as string;
@@ -78,15 +83,12 @@ async function getUser(req: NextApiRequest): Promise<Array<User>>  {
     const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
     return emailRegex.test(email);
   };
-  const email = req.query.email as string;
-  const password = req.query.password as string;
+  const email = req.query.id as string;
   const isValid = validateEmail(email)
-
-  if (isValid) {
+    if (isValid) {
     return await prisma.user.findMany({
       where: {
         email: email,
-        password:  password
       }, 
       select: {
         id: true,
